+++ tools/PSc8y/Public-manual/Get-C8ySessionProperty.ps1
@@ -27,13 +27,17 @@ string
         [string] $Name
     )
 
-    switch ($Name) {
-        "tenant" {
-            $env:C8Y_TENANT
-        }
-
-        "host" {
-            $env:C8Y_HOST
-        }
+    $Values = @{
+        host = $env:C8Y_HOST
+        tenant = $env:C8Y_TENANT
     }
+
+    # Is calling cmdlet using the session variable?
+    $SessionFile = (Get-PSCallStack).InvocationInfo.BoundParameters.Session
+
+    if ($SessionFile) {
+        $Values = Get-Content -LiteralPath $SessionFile | ConvertFrom-Json
+    }
+
+    $Values[$Name]
 }