+++ tools/PSc8y/Public-manual/Get-SessionCollection.ps1
@@ -18,9 +18,14 @@ object[]
     Param()
     $HomePath = Get-SessionHomePath
 
+    $JSONArgs = @{}
+    if ($PSVersionTable.PSVersion.Major -gt 5) {
+        $JSONArgs.Depth = 100
+    }
+
     $Sessions = Get-ChildItem -LiteralPath $HomePath -Filter "*.json" -Recurse | ForEach-Object {
         $Path = $PSItem.FullName
-        $data = Get-Content -LiteralPath $Path | ConvertFrom-Json
+        $data = Get-Content -LiteralPath $Path | ConvertFrom-Json @JSONArgs
         $data | Add-Member -MemberType NoteProperty -Name "path" -Value $Path -ErrorAction SilentlyContinue
         $data.path = $Path
         $data